+++ forest4/area2/index.html
@@ -4,9 +4,9 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-  <link rel="stylesheet" href="./styles/main.css">
-  <link rel="stylesheet" href="./styles/animation.css">
-  <script type="module" src ="./scripts/area/app2.js" charset="utf-8"></script>
+  <link rel="stylesheet" href="../styles/main.css">
+  <link rel="stylesheet" href="../styles/animation.css">
+  <script type="module" src ="app.js" charset="utf-8"></script>
 
   <title>testFloor</title>
 </head>
