+++ forest4/index.html
@@ -6,7 +6,7 @@
 
   <link rel="stylesheet" href="./styles/main.css">
   <link rel="stylesheet" href="./styles/animation.css">
-  <script type="module" src ="./scripts/area/app.js" charset="utf-8"></script>
+  <script type="module" src ="./area1/app.js" charset="utf-8"></script>
 
   <title>forest4</title>
 </head>
